feat: Enhance subscription and external squad management
- Updated NodeUsageDto to include node_uuid and changed date to datetime.
- Changed total fields in subscription and subscription request history models from float to int.
- Introduced response rules and conditions for subscription settings, including new models for response modifications.
- Added external squads management with CRUD operations and associated models.
- Implemented passkey management with registration and verification endpoints.
- Enhanced snippets management with full CRUD operations and validation.
- Added OAuth2 provider enum for better authentication handling.

+++ remnawave/models/inbounds.py
@@ -17,7 +17,7 @@ class InboundResponseDto(BaseModel):
 
 
 class AllInboundsData(BaseModel):
-    total: float
+    total: int
     inbounds: List[InboundResponseDto]
 
 
@@ -26,7 +26,7 @@ class GetAllInboundsResponseDto(AllInboundsData):
 
 
 class InboundsByProfileData(BaseModel):
-    total: float
+    total: int
     inbounds: List[InboundResponseDto]
